(feat) moved classHelpers into modelHelpers
