+++ src/server/db/ModelHelpers/classHelpers.js
